Flask_blog/backend/requirements.txt
flask==3.0.3
flask_sqlalchemy==3.1.1
flask_migrate==4.0.5
flask_bcrypt==1.0.1
flask_cors==4.0.0
PyJWT==2.8.0
python-dotenv==1.0.1
redis==5.0.7
celery==5.4.0
meilisearch==0.31.4
pillow==10.4.0
bleach==6.1.0
markdown==3.6
sqlalchemy==2.0.32
alembic==1.13.2
requests==2.32.3
python-slugify==8.0.4
email-validator==2.2.0
pytest==8.3.2
flask-limiter==3.8.0
# 生产部署 & MySQL
gunicorn==21.2.0
# MySQL drivers (选择其一即可，默认使用 PyMySQL)
PyMySQL==1.1.1
mysqlclient==2.2.4
# 请求验证
pydantic==2.8.2
# 指标监控
prometheus-client==0.20.0
# 国际化
yapf==0.40.2
Flask-Babel==4.0.0
